Add GetLocalPendingMembersWithInfo method
Add GetLocalPendingMembersWithInfo method to the group interface. Rationale is
that the local pending members represent requests to the client. To be able to
make a decision about a request the client really needs to get all Info
container the request.. e.g whom invited the client to join a group-chat and
why..


20070201151745-93b9a-290c315ea9f1b07fdbe88f7a09a118f62b8c46e0.gz