Emit Text.Sent, Text.Received even for non-text messages on Channel.Interface.MessageParts.
This means that Text-only clients can acknowledge messages received with no
text content (they will probably appear empty, but this is about the best
we can do - if you've received the non-textual message, then capability
advertisement has already failed).

This avoids a problem with the previous draft, where non-Text messages
(messages on the MessageParts interface that have no text/plain content and so
were not signalled on the Text channel) could never be acknowledged or removed
by Text-only clients, and would build up in memory indefinitely.

20080410191611-53eee-abf5ff7f56ddc7e42046eee0bcf9058bd9cb89cc.gz