Message_Part: add array-depth
Unlike telepathy-glib, telepathy-qt *needs* named types for (almost)
every type used in generated code. In this case we need to have a name
for an array of arrays of mappings (aaa{sv}). array²-name was proposed,
but is not well-formed XML :-( and doesn't scale if we ever need
an array of arrays of arrays of mappings (etc.).

The semantics of array-depth are:
* if array-depth is omitted, its value defaults to 1 if there is an
  array-name, or 0 otherwise
* if array-depth >= 1, an array-name must be given
* a number of array types equal to array-depth is generated
* the first array type is named using array-name
* subsequent array types are named in an implementation-dependent way:
  currently TelepathyQt4 just appends "List" repeatedly

Signed-off-by: Simon McVittie <simon.mcvittie@collabora.co.uk>