Rename member-ids to contact-ids and clarify that it's only an optimization
The contacts aren't necessarily all members of the group (in XMPP you
can be kicked by someone not currently in the chatroom) so member-ids was
a misleading name.

In addition, we explicitly allow the CM to include and exclude arbitrary
handle/ID pairs in the mapping, or omit it entirely; clients must have a
fallback path anyway, so this doesn't make clients any more complicated.