MediaSignalling: remove "lowering barrier to entry" wording, and specify that transports passed to EnsureChannel etc. are ignored
The "lowering barrier to entry" wording would allow installation of a
client A (that handles ICE only) to break an unrelated client B's support
for GTalkP2P, if B relied on this mechanism. As a result, nobody can
rely on this mechanism, so it's basically useless.

The ability to disable transports by parameters to EnsureChannel etc. is
only a partial solution, since channels requested by an address book
can't take advantage of this mechanism. A better solution (to be added
later) would be to introduce interactive transport negotiation, done by
the streaming implementation (e.g. before Ready).