Remove x-search-offset.
This should have been a requestable property, not a pseudo-term. Also,
it's not very useful: you can't be sure that a second search with a
higher offset will actually be indexing the same list of results. If we
want to support protocols which do support paged retrieval of results,
we could add a Paged property, whose value is the number of items per
page, and a NextPage() method, callable while the search is In_Progress,
to ask for the next page. Or something.