fd.o #21089: Observer: clarify behaviour when some, but not all, channels in a batch are interesting
Imagine that Empathy observes text and streamed-media channels, and Kopete
observes text channels only. A NewChannels signal announces a text channel and
a streamed-media channel simultaneously; obviously, Empathy gets
ObserveChannels called with both channels as arguments.

telepathy-spec did not previously specify whether Kopete's ObserveChannels
method would be called with the text channel, or with both channels.
In Mission Control 5 it only gets the text channel - i.e. the subset of
the bundle that it understands - which seems like reasonable behaviour.