Include immutable properties in Connection.Sidecars
a{oa{sv}} is easier to cope with in most bindings than a(oa{sv}), and
allows the property to include the sidecars' immutable properties as
well as which interfaces they implement.

Additionally, define the lifetime of each sidecar, and when the property
can change (once, when moving to Connected).