Make Connection.Sidecars an a{s(oa{sv})}
Discussion with the spec cabal reached the conclusion that the thing
clients care about most is the main interface a sidecar implements.
Looking up an interface in a map is much easier than iterating a map
from object paths to details, looking up o.fd.T.Sidecar.Interfaces in
the details, and checking whether the interface you care about is in
that list.