Replace the Sidecars property with EnsureSidecar()
While sketching the implementation in Gabble, we noticed that change
notification on moving to state Connected is inadequate: connecting to
some server component might take a long time, and you don't want the
rest of the connection to wait on that before being useful.

But a SidecarAdded signal is not sufficient: clients need to know
whether a given sidecar is ever going to appear. So it seems sensible to
add a method to get a given sidecar, which can take a long time to
return and yield an error if necessary. Then, the Sidecars property
becomes redundant: either you know which sidecars you want (in which
case you can ask for them explicitly), or you don't (in which case, what
would you do with the information?).