Add a signal to indicate if the nat server info has been retrieved
Farsight currently needs both the stun and relay info to be set before it can
start doing anything with a stream. Hopefully in the future this will change.
But for now having seperate StunServerAdded and RelayInfoAdded signals are
pointless.

When farsight is eventually updated to be able to add stun and relay servers on
the fly, this signal is still needed as we need to be able to tell some
protocols that we have an initial set of candidates, which needs to wait for an
initial set of stun server and relay servers to have been retrieved.