MailNotification: rename Capabilities to MailNotificationFlags
In theory, D-Bus is nicely namespaced. In practice,  bindings don't
fully support namespacing (in particular, dbus-glib and QtDBus will be
unhappy if you have a property of the same name in different interfaces),
so we should rename Capabilities to something less generic.

I named it to match its type, but MailNotificationSupport would also
be a reasonable name if you prefer that.