AddClientInterest: allow interests to be tokens starting with an interface
All of them that we've thought of so far are interface names, but it seems
likely that we'll want to have more than one interest on an interface at
some point.

Hypothetical use-case: you could subscribe to ...MailNotification/count
if you only care about the number of messages, not the details.