Fix rationale for CSHs on GSM conferences
Actually, the old rationale was somewhat true: as I understand it, you
may get a notification when the remote party merges you into a
conference, and perhaps when they add or remove others from the
conference, but you're not guaranteed to get all of these notifications
(particularly not the downwards edges). If we want to expose this later,
it should live on a new interface.