Specify GSM calls becoming frozen and unfrozen.
GSM doesn't require you to put both calls on hold — that's not even
possible, you can only have one held call at a time.

I think we're best leaving this quirk of GSM as a quirk of GSM,
specified in the specification. We could later add some API for telling
the difference, or what exactly each CM should fake, if we find we need
it.