Add a tp:is-connection-parameter property annotation
We have an increasing number of standard connection properties which are
also connection parameters with the DBus_Property flag. They all tend to
have some copy-pasta blurb explaining that clients should basically
ignore the property and frob the parameter on the Account instead; I
think it's time to add some markup for this.