Fix wording in [sometimes] requestable templates
The sometimes template said "this property requestable which" where it
meant "this property may be requestable, which". Also, both templates
said "such as in the\nthe".

Also "Fixed_Properties or Allowed_Properties lists" suggested that both
those struct fields are lists, whereas the former is a hash. So I binned
the "lists".