Add a Base.get_anchor() method, to be used in HTML anchors.
Update get_url() to use get_anchor().

This removes the duplicate construction of anchors, once in the templates
and once in the get_url() method and move it into one place. This allows us
to easily change the way anchors are constructed for a given type.

The way anchors are constructed has been changed to include the type name
at the beginning, i.e. Type-Name:Short_Name, to prevent potential namespace
collisions between client interests and capability tokens.