Replace CodecOffer by Description objects
For several advanced features we want to Offer more then just the codecs to the
local side. This changes the interface away from a very codec-centric way and
towards using a more generic description of the remote (and local offers).

The other big change is that we now track our local offers per-handle. This is
so that non-codec information that can be different per remote-end can be
tracked explicitely over D-Bus.