Makefile: clarify check and dist rules
Currently large swathes of 'check' and 'dist' are a series of shell
commands chained together into one command for make's purposes, with set
-e to make the first error fatal. This makes it hard to see what's wrong
when they fail. For instance, the case I hit today was that I'd
forgotten to sign the tarball; I had no way of knowing that it was the
`test -f telepathy-spec-$$version.tar.gz.asc` command that had failed.

In some realities the right solution would be to import
maintainer-make-release from tp-glib, but I don't really fancy
disentangling the Automakey bits from that right now. So this patch
basically just breaks the über-commands into individual commands.

(This syntax:

    FOO := bar

is a GNU Make-ism which means “evaluate bar once, and then assign its
result to FOO”. If we instead used:

    VERSION = `sed ...`

then the sed invocation would be evaluated once for each use of
$(VERSION). Since we already depend on GNU Make for things like
$(filter-out) I see no reason not to use the eminently-sensible :=
syntax.)