Remove Call_State_Pending_Receiver
(It is covered by its "substates". I also realised that the concept of
substates is bullshit and confusing in an enum)

I have specified that Ringing should be set immediately on protocols that
don't signal Ringing as a state. This may be slightly controversial, since
it's a loss of information, but I think this reflects what a UI should do
in this case.

* Grep for Pending_Receiver and fix all instances
* Update the enum's numbers, and remove all references to "sub-state"
* Remove "makes sense on calls in state..." from Call_Flags, because
  they all kinda make sense in all states apart from Ended.