Style tables in docstrings more pleasingly
The colours and styles are designed to match the summary at the start of
each interface page.

The tr:nth-child(even) rule styles every other row with a slightly
darker background, which makes it easier to read the only two tables we
have (the well-known status identifiers, and the many states of Call
channels). :nth-child is from CSS3. It looks fine without it, but better
with it.

border-collapse: collapse; is like cellspacing='0'.