Revert "Make CM responsible for on-disk avatar caching"
This is going to make it really difficult to port existing CMs to
Telepathy 1.0. Let's reduce the amount of rewriting we have to do
in the initial port.

When we make this change (either for 1.0 or 1.2), we should add
Avatars1 to master, do a mixin which encapsulates it, and port CMs
to that mixin.

This reverts commit 1f972813ca8cb53382534085263138805bacca19.