Call Renaming a stable API
The current version is identical to what's in telepathy-idle, which
has had it without implementation changes since 2007. If that's
not stable API, nothing is :-)

I deliberately haven't called it Renaming1 (although I will in the
'next' branch), so that in the unlikely event that any UI pays
attention to it, it won't need changes. Not only is Renaming older
than the Foo1 convention, it's also older than the Foo.DRAFT
convention, so it never had either.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=25147
Reviewed-by: Guillaume Desmottes <guillaume.desmottes@collabora.co.uk>