Import ChannelDispatcher.I.Messages from Mission Control master
My main concern about this interface was whether implementing it
caused MC to be unmaintainable, but it isn't too bad in practice,
just unfinished.

The one edit here, relative to Mission Control, is making the
filename end with "1.xml" instead of "_DRAFT.xml", so that
subsequent diffs on the contents aren't lost in the rename noise.

Bug: https://bugs.freedesktop.org/show_bug.cgi?id=37380
Reviewed-by: Guillaume Desmottes <guillaume.desmottes@collabora.co.uk>